+++ camply-web/src/features/academic-form/components/NavigationControls.tsx
