+++ camply-web/src/features/academic-form/components/ProgressIndicator.tsx
