+++ camply-web/src/features/academic-form/components/steps/DepartmentStep.tsx
